# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CubeSandbox Agent Adapter API
  version: 0.3.0
  description: >-
    Authenticated, tenant-aware execution broker for CubeSandbox microVMs.
    Full Sandbox IDs, Cube traffic tokens, commands, output, and bearer tokens
    are intentionally excluded from audit records.
servers:
  - url: http://127.0.0.1:18080
security:
  - bearerAuth: []
paths:
...
  /v1/leases/acquire:
    post:
      operationId: acquireLease
      requestBody: { $ref: "#/components/requestBodies/Acquire" }
      responses:
        "200": { $ref: "#/components/responses/JsonSuccess" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Denied" }
        "429": { $ref: "#/components/responses/Error" }
...
  /v1/leases/{lease_ref}/exec:
    post:
      operationId: executeCommand
      parameters: [{ $ref: "#/components/parameters/LeaseRef" }]
      requestBody: { $ref: "#/components/requestBodies/Exec" }
      responses:
        "200": { $ref: "#/components/responses/JsonSuccess" }
        "404": { $ref: "#/components/responses/NotFound" }
        "504": { $ref: "#/components/responses/Error" }
  /v1/leases/{lease_ref}/read:
    post:
      operationId: readTextFile
      parameters: [{ $ref: "#/components/parameters/LeaseRef" }]
      requestBody: { $ref: "#/components/requestBodies/Path" }
      responses:
        "200": { $ref: "#/components/responses/JsonSuccess" }
        "403": { $ref: "#/components/responses/Denied" }
  /v1/leases/{lease_ref}/write:
    post:
      operationId: writeTextFile
      parameters: [{ $ref: "#/components/parameters/LeaseRef" }]
      requestBody: { $ref: "#/components/requestBodies/Write" }
      responses:
        "200": { $ref: "#/components/responses/JsonSuccess" }
        "403": { $ref: "#/components/responses/Denied" }
...
  /v1/leases/{lease_ref}/release:
    post:
      operationId: releaseLease
      parameters: [{ $ref: "#/components/parameters/LeaseRef" }]
      requestBody: { $ref: "#/components/requestBodies/Release" }
      responses:
        "200": { $ref: "#/components/responses/JsonSuccess" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-JWT
  parameters:
    LeaseRef:
      name: lease_ref
      in: path
      required: true
      schema: { type: string, pattern: "^lease_[a-f0-9]{20}$" }
...
  requestBodies:
    Acquire:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AcquireRequest" }
    Exec:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ExecRequest" }
    Path:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PathRequest" }
    Write:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/WriteRequest" }
...
    Release:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ReleaseRequest" }
...
  responses:
    JsonSuccess:
      description: Successful JSON response
      content:
        application/json:
          schema: { type: object }
...
    Error:
      description: Adapter error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Denied:
      description: Authenticated caller is not permitted to perform the operation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource is absent or hidden by tenant isolation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    RequestId:
      type: string
      pattern: "^[A-Za-z0-9_.:-]{1,80}$"
    Path:
      type: string
      pattern: "^/(workspace|tmp)(/|$)"
...
    AcquireRequest:
      type: object
      additionalProperties: false
      required: [runtime, session_key]
      properties:
        runtime: { type: string, enum: [openclaw, dsh, hermes, mcp] }
        session_key: { type: string, minLength: 1, maxLength: 512 }
        profile: { type: string, default: offline-code }
        request_id: { $ref: "#/components/schemas/RequestId" }
    ExecRequest:
      type: object
      additionalProperties: false
      required: [command]
      properties:
        command: { type: string, minLength: 1, maxLength: 16384 }
        cwd: { $ref: "#/components/schemas/Path" }
        timeout_ms: { type: integer, minimum: 1 }
        request_id: { $ref: "#/components/schemas/RequestId" }
    PathRequest:
      type: object
      additionalProperties: false
      required: [path]
      properties:
        path: { $ref: "#/components/schemas/Path" }
        request_id: { $ref: "#/components/schemas/RequestId" }
    WriteRequest:
      type: object
      additionalProperties: false
      required: [path, content]
      properties:
        path: { $ref: "#/components/schemas/Path" }
        content: { type: string, maxLength: 262144 }
        request_id: { $ref: "#/components/schemas/RequestId" }
...
    ReleaseRequest:
      type: object
      additionalProperties: false
      properties:
        action: { type: string, enum: [pause, kill], default: pause }
        request_id: { $ref: "#/components/schemas/RequestId" }
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }